Update fPIC and shared lib configuration
Conan has some built in options for 'fPIC', which sets
CMAKE_POSITION_INDEPENDENT_CODE ON, and 'shared', which sets
BUILD_SHARED_LIBS.
I implemented these options in the conanfile, and set it up so that the
project respects these options

I also set it up so that if this library is used purely with cmake with add_subdirectory or fetchContent, the library will respect CMAKE_POSITION_INDEPENDENT_CODE and BUILD_SHARED_LIBS of the consumer.
If the consumer does not want to set these options globally, a variable for building shared libs and using fPIC for the static lib just for this project are present.

I also updated the build_docs and build_tests conan options to not affect the package id, since they do not affect the library binary.